\name{ggobi.scagdf}
\alias{ggobi.scagdf}
\title{Open ggobi with matching R plot}
\author{Hadley Wickham <h.wickham@gmail.com>}

\description{
Experimental interface to interact with scagnostics output with GGobi
}
\usage{ggobi.scagdf(data, ...)}
\arguments{
\item{data}{}
\item{...}{}
}

\details{This is an experimental interface to explore the scagnostics output
with GGobi, taking advantage of the fact that each point in the output
of \code{\link{scagnostics}} corresponds to a plot.  To use it, switch into
identify mode and hover over the plot (= point) that you want to see.}

\examples{}
\keyword{dynamic}
